ingester.v1.bak/requirements.txt
psycopg2-binary
requests
redis
praw
prometheus-client
